2025/braintrust-otel/requirements.txt
fastapi[standard]==0.115.12 
openai==1.78.0
opentelemetry-exporter-otlp-proto-http==1.33.0
opentelemetry-instrumentation-fastapi==0.54b0 
opentelemetry-instrumentation-openai==0.40.4
opentelemetry-sdk==1.33.0
python-dotenv==1.1.0
